Update features.csv
I have updated the number of datasets and weights in TerraTorch. 
We don’t have the concept of datasets in TerraTorch, we rely on TorchGeo for that, so I changed the count of datasets to 0 put in parentheses the same number as TorchGeo. 
As for weights, we currently have integrated 1374 models into TerraTorch (1282 are TIMM models). The count can be obtained by running: len(BACKBONE_REGISTRY) in TerraTorch.